HDDS-2476. Share code between ContainerMetadataScrubberMetrics and ContainerDataScrubberMetrics
